+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Bug fixes
 - Ensure that only non-redirect URL rewrites are considered when generating product URLs - thank you @fasimana
+- Apply rounding to insight events revenue values to avoid floating point precision errors - thank you @PromInc 
 - Fix store id for queue jobs merging and sorting - thank you @pikulsky
 
 ## 3.15.1